+++ Editor/AssemblyInfo.cs
@@ -2,3 +2,5 @@
 
 [assembly: InternalsVisibleTo("Unity.VisualStudio.EditorTests")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
+[assembly: InternalsVisibleTo("Hackerzhuli.Code.Entities")]
+[assembly: InternalsVisibleTo("Hackerzhuli.Code.Entities.Tests")]